removed many unused classes, funcionalities, dependencies. Made huge impact on changing map to roadmap and marsker instead of infoView
