--- src/main/java/allinone/Config.java
@@ -1,17 +1,0 @@
-//package allinone;
-//
-//import org.springframework.context.annotation.Configuration;
-//import org.springframework.core.Ordered;
-//import org.springframework.core.annotation.Order;
-//import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-//import org.springframework.security.config.annotation.authentication.configurers.GlobalAuthenticationConfigurerAdapter;
-//
-////@Order ( Ordered.HIGHEST_PRECEDENCE)
-////@Configuration
-////public class Config extends GlobalAuthenticationConfigurerAdapter {
-////
-////	@Override
-////	public void init(AuthenticationManagerBuilder auth) throws Exception {
-////		System.err.println("INIT in Authentication manager ! ");
-////	}
-////}
