--- src/main/java/allinone/SecurityConfiguration.java
@@ -1,43 +1,0 @@
-package allinone;
-
-import org.springframework.boot.autoconfigure.security.SecurityProperties;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.annotation.Order;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.annotation.web.servlet.configuration.EnableWebMvcSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-
-
-@Configuration
-@EnableWebMvcSecurity
-@Order(SecurityProperties.ACCESS_OVERRIDE_ORDER)
-class SecurityConfiguration extends WebSecurityConfigurerAdapter {
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-
-//        http.authorizeRequests().antMatchers("/login").permitAll()
-//                .antMatchers("/**", "/all", "/console/**", "/login/**", "/css/**", "/images/**", "/js/**", "/fonts/**")
-//                .permitAll()
-//                .anyRequest()
-//                .authenticated()
-//                .and().exceptionHandling()
-//                .accessDeniedPage("/login?error")
-//
-//        ;
-
-        http
-                .sessionManagement()
-                .maximumSessions(1)
-                .expiredUrl("/login?expired")
-                .maxSessionsPreventsLogin(true)
-                .and()
-                .sessionCreationPolicy(SessionCreationPolicy.ALWAYS)
-                .invalidSessionUrl("/");
-
-
-    }
-
-
-}
