--- src/main/java/allinone/SessionListener.java
@@ -1,18 +1,0 @@
-package allinone;
-
-import javax.servlet.http.HttpSessionEvent;
-import javax.servlet.http.HttpSessionListener;
-
-public class SessionListener implements HttpSessionListener {
-
-    @Override
-    public void sessionCreated(HttpSessionEvent event) {
-        System.out.println("session created, set interval");
-        event.getSession().setMaxInactiveInterval(10);
-    }
-
-    @Override
-    public void sessionDestroyed(HttpSessionEvent event) {
-        System.out.println("session destroyed");
-    }
-}
