--- src/main/java/allinone/WebSocketConfig.java
@@ -1,56 +1,0 @@
-package allinone;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.messaging.MessagingException;
-import org.springframework.messaging.handler.annotation.DestinationVariable;
-import org.springframework.messaging.handler.annotation.MessageMapping;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
-import org.springframework.messaging.simp.annotation.SendToUser;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.scheduling.TaskScheduler;
-import org.springframework.scheduling.annotation.EnableScheduling;
-import org.springframework.scheduling.concurrent.ConcurrentTaskScheduler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.socket.config.annotation.AbstractWebSocketMessageBrokerConfigurer;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-
-@Configuration
-@EnableWebSocketMessageBroker
-@EnableScheduling
-@Component
-public class WebSocketConfig extends AbstractWebSocketMessageBrokerConfigurer {
-
-	@Override
-	public void configureMessageBroker(MessageBrokerRegistry config) {
-		config.enableSimpleBroker("/topic");
-		config.setApplicationDestinationPrefixes("/app");
-	}
-
-
-	@Override
-	public void registerStompEndpoints(StompEndpointRegistry registry) {
-		registry.addEndpoint("/hello").withSockJS();
-		registry.addEndpoint("/topic/greetings").withSockJS();
-		registry.addEndpoint("/ws/endpoint").withSockJS();
-		registry.addEndpoint("/topic/greetings/{groupID}").withSockJS();
-
-	}
-
-    @MessageMapping("/hello/{groupID}")
-    @SendToUser("/topic/greetings")
-    public String greeting(
-//            SimpMessagingTemplate template,
-    @PathVariable String groupID
-    ){
-		System.out.println("trolololo " + groupID);
-        //create
-
-
-		return "Test Sended String";
-    }
-    
-    
-}
