--- src/main/java/allinone/repositories/GroupRepository.java
@@ -1,16 +1,0 @@
-package allinone.repositories;
-
-
-import allinone.entities.Group;
-import allinone.entities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import javax.transaction.Transactional;
-
-@Repository
-@Transactional
-public interface GroupRepository extends JpaRepository<Group, Long> {
-	Group findById(String id);
-}
-
-
