--- src/main/java/allinone/repositories/UserRepository.java
@@ -1,16 +1,0 @@
-package allinone.repositories;
-
-
-import allinone.entities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import javax.transaction.Transactional;
-
-@Repository
-@Transactional
-public interface UserRepository extends JpaRepository<User, Long> {
-	User findById(String id);
-}
-
-
