--- src/main/java/allinone/service/LocationService.java
@@ -1,24 +1,0 @@
-package allinone.service;
-
-import org.springframework.stereotype.Service;
-
-/**
- * Created by Kowcio on 2016-10-08.
- */
-
-@Service
-public class LocationService {
-
-
-
-    ///mocked stuff
-    String GEO_API = "AIzaSyBXY_QpdRRRbMQ3yGKO8Wc_fYse9xVP2TE";
-    String city = "city";
-    String longitude;
-    String latitude;
-
-
-
-    
-
-}
