--- src/main/java/allinone/web/controller/RestController.java
@@ -1,46 +1,0 @@
-package allinone.web.controller;
-
-import allinone.service.GroupService;
-import allinone.service.LocationService;
-import allinone.service.UserService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.servlet.ModelAndView;
-
-@Controller
-@ResponseBody
-@RequestMapping(value = "/api/", method = RequestMethod.GET)
-
-public class RestController {
-    
-    private final Logger log = LoggerFactory.getLogger(getClass());
-
-
-    @Autowired
-    GroupService groupService;
-    @Autowired
-    LocationService locationService;
-    @Autowired
-    UserService userService;
-
-
-    @RequestMapping(value = "/newgroup/", method = RequestMethod.GET)
-    public ModelAndView addGroup() {
-
-        ModelAndView mav = new ModelAndView("/pages/main");
-        mav.addObject("User", userService.getRandomUser());
-
-        return mav;
-    }
-
-
-
-
-
-
-}
