--- src/main/java/allinone/web/controller/utils/CONST.java
@@ -1,10 +1,0 @@
-package allinone.web.controller.utils;
-
-/**
- * Class with the collection of  reusable strings in the system.
- */
-public class CONST {
-
-    String MSG = "Test";
-
-}
