+++ src/main/java/karawana/Application.java
@@ -1,0 +1,44 @@
+package karawana;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cache.interceptor.KeyGenerator;
+import org.springframework.cache.interceptor.SimpleKeyGenerator;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+
+@SpringBootApplication
+/*
+ * 4 in 1
+ *
+ * @Configuration
+ *
+ * @EnableAutoConfiguration
+ *
+ * @EnableWebMvc
+ *
+ * @ComponentScan
+ */
+@EnableJpaAuditing
+@EnableScheduling
+@EnableCaching
+@EnableTransactionManagement
+public class Application extends WebMvcConfigurerAdapter {
+    final static Logger log = LoggerFactory.getLogger(Application.class);
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+
+    @Bean
+    public KeyGenerator keyGenerator() {
+        return new SimpleKeyGenerator();
+    }
+
+}
