+++ src/main/java/karawana/config/MyCacheManager.java
@@ -1,0 +1,27 @@
+package karawana.config;
+
+import org.springframework.cache.Cache;
+import org.springframework.cache.CacheManager;
+import org.springframework.cache.concurrent.ConcurrentMapCache;
+import org.springframework.cache.support.SimpleCacheManager;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Created by U6041021 on 02/02/2017.
+ */
+@Configuration
+public class MyCacheManager {
+
+    @Bean
+    public CacheManager cacheManager() {
+        SimpleCacheManager cacheManager = new SimpleCacheManager();
+        List<Cache> caches = new ArrayList<>();
+        caches.add(new ConcurrentMapCache("myCache"));
+        cacheManager.setCaches(caches);
+        return cacheManager;
+    }
+}
