+++ src/main/java/karawana/config/SessionListener.java
@@ -1,0 +1,18 @@
+package karawana.config;
+
+import javax.servlet.http.HttpSessionEvent;
+import javax.servlet.http.HttpSessionListener;
+
+public class SessionListener implements HttpSessionListener {
+
+    @Override
+    public void sessionCreated(HttpSessionEvent event) {
+        System.out.println("session created, set interval");
+        event.getSession().setMaxInactiveInterval(10);
+    }
+
+    @Override
+    public void sessionDestroyed(HttpSessionEvent event) {
+        System.out.println("session destroyed");
+    }
+}
