+++ src/main/java/karawana/entities/Location.java
@@ -1,0 +1,26 @@
+package karawana.entities;
+
+import lombok.Builder;
+import lombok.Data;
+
+import javax.persistence.*;
+
+
+@Data
+@Builder
+@Entity
+public class Location //extends AbstractEntity
+{
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    protected Long id = 0L;
+
+    @Version
+    private Long version;
+
+    private Double latitude;
+    private Double longitude;
+
+
+
+}
