+++ src/main/java/karawana/repositories/LocationRepository.java
@@ -1,0 +1,17 @@
+package karawana.repositories;
+
+
+import karawana.entities.Location;
+import karawana.entities.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import javax.transaction.Transactional;
+
+@Repository
+@Transactional
+public interface LocationRepository extends JpaRepository<Location, Long> {
+	User findById(String name);
+}
+
+
