+++ src/main/java/karawana/repositories/UserRepository.java
@@ -1,0 +1,16 @@
+package karawana.repositories;
+
+
+import karawana.entities.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import javax.transaction.Transactional;
+
+@Repository
+@Transactional
+public interface UserRepository extends JpaRepository<User, Long> {
+	User findById(String id);
+}
+
+
