+++ src/main/java/karawana/service/LocationService.java
@@ -1,0 +1,21 @@
+package karawana.service;
+
+import org.springframework.stereotype.Service;
+
+/**
+ * Created by Kowcio on 2016-10-08.
+ */
+
+@Service
+public class LocationService {
+    ///mocked stuff
+    String GEO_API = "AIzaSyBXY_QpdRRRbMQ3yGKO8Wc_fYse9xVP2TE";
+    String city = "city";
+    String longitude;
+    String latitude;
+
+
+
+    
+
+}
