+++ src/main/resources/application.properties
@@ -1,61 +1,26 @@
-#REF
-#http://docs.spring.io/spring-boot/docs/1.2.x-SNAPSHOT/reference/htmlsingle/#common-application-properties
+#Websockets
+#allinone.service.WebSocketBroadcastService.isOn = false
 
-
-# MY defined global vars
-
-allinone.service.WebSocketBroadcastService.isOn = true
-
-
-
-#   override catalina 
-tomcat.util.buf.StringCache.byte.enabled=true
-tomcat.util.buf.StringCache.char.enabled=false
-tomcat.util.buf.StringCache.trainThreshold=50000
-tomcat.util.buf.StringCache.cacheSize=5000
-
-
-
+#H2-Base properties
+spring.h2.console.enabled=true
+spring.h2.console.path=/console
+spring.datasource.platform=h2
 spring.datasource.initialize=true
 
-#logging.level.org.springframework.security: INFO
-#logging.level.org.springframework.jdbc: INFO
-#logging.level.org.springframework.web: INFO
-#logging.level.org.hibernate: INFO
-#logging.level.com.googlecode.flyway.core.dbsupport.SqlScript: INFO
-
+#JPA
 spring.main.show-banner=false
-
-#spring.datasource.data=/db/migration/common/V1__add.sql 
-#spring.jpa.generate-ddl=true
-#overrides the above one
+spring.data.jpa.repositories.enabled=true
 spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.hibernate.dialect=org.hibernate.dialect.H2Dialect
 spring.jpa.hibernate.format_sql=true
 spring.jpa.show-sql=true
-spring.data.jpa.repositories.enabled=true
-#https://stackoverflow.com/questions/8974747/hibernate-4-and-joda-time
-#jadira.usertype.autoRegisterUserTypes=true
-#jadira.usertype.databaseZone=jvm
-#jadira.usertype.javaZone=jvm
 
+#Logging properties
+logging.level.org.springframework.security: INFO
+#logging.level.org.springframework.jdbc: INFO
+#logging.level.org.springframework.web: INFO
+logging.level.org.hibernate: INFO
+logging.level.com.googlecode.flyway.core.dbsupport.SqlScript: INFO
 
-spring.thymeleaf.cache= false
-#spring.datasource.schema=/resources/sql/schema-*.sql
-#security.basic.enabled=false
-
-test = "This is test string value from application.properties file."
-# FLYWAY
-
-flyway.locations=db/migration/common
-flyway.sql-migration-prefix=V
-flyway.sql-migration-suffix=.sql
-flyway.driver=org.hibernate.dialect.H2Dialect
-flyway.url=jdbc:h2:./skeledb;AUTO_SERVER=FALSE;IFEXISTS=FALSE;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;MVCC=FALSE
-
-# INTERNATIONALIZATION (MessageSourceAutoConfiguration)
-#spring.messages.basename=messages
-#spring.messages.cacheSeconds=-1    
-#spring.messages.encoding=UTF-8
