--- src/main/resources/logback.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<configuration debug="true">
-
-	<appender name="STDOUT" class="ch.qos.logback.core.ConsoleAppender">
-		<encoder>
-		
-			<pattern>%n%d{HH:mm:ss,SSS} #  [TH:%20t] %p # %10C{1}.%.20M{1}.%L ### %m # %r ms)</pattern>
-			
-		</encoder>
-	</appender>
-
-	<root level="info">
-		<appender-ref ref="STDOUT" />
-	</root>
-</configuration>
