--- src/main/resources/rebel.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<application xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://www.zeroturnaround.com" xsi:schemaLocation="http://www.zeroturnaround.com http://www.zeroturnaround.com/alderaan/rebel-2_0.xsd">
-
-	<classpath>
-		<dir name="C:/Users/Kowcio/git/allinone/target/classes">
-		</dir>
-	</classpath>
-
-</application>
